Renamed a permission file.
