+++ token_manager/pages/index.tsx
@@ -46,7 +46,7 @@ import { findValue, getBasicPolyWalletApi, returnUpdatedCreator } from "../src/u
 import { CheckpointView } from "../src/components/checkpoints/CheckpointView"
 import { CheckpointScheduleView } from "../src/components/checkpoints/CheckpointScheduleView"
 import { CheckpointManagerView } from "../src/components/checkpoints/CheckpointManagerView"
-import { PermissionManagerView } from "../src/components/permissions/PermissionView"
+import { PermissionManagerView } from "../src/components/permissions/PermissionManagerView"
 import { ComplianceManagerView } from "../src/components/compliance/ComplianceView"
 import { LongHexView } from "../src/components/LongHexView"
 import { PortfoliosView, PortfolioView } from "../src/components/portfolios/PortfolioView"
