+++ token_manager/src/components/permissions/PermissionManagerView.tsx
